+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+Upcoming Version (WIP)
+=================
+
 8.6.0 / 2026-02-05
 =================
 Improvements: